+++ src/irvine/oeis/a006/A006129.java
@@ -17,7 +17,7 @@ public class A006129 extends Sequence0 {
     ++mN;
     Z sum = Z.ZERO;
     for (int k = 0; k <= mN; ++k) {
-      sum = sum.signedAdd(((mN - k) & 1) == 0, Binomial.binomial(mN, k).shiftLeft(k * (k - 1) / 2));
+      sum = sum.signedAdd(((mN - k) & 1) == 0, Binomial.binomial(mN, k).shiftLeft((long) k * (k - 1) / 2));
     }
     return sum;
   }
