+++ src/irvine/oeis/a006/A006312.java
@@ -26,7 +26,7 @@ public class A006312 extends A006308 {
   @Override
   public Z next() {
     mP = (int) mPrime.nextPrime(mP);
-    final long p2 = mP * mP;
+    final long p2 = (long) mP * mP;
     final Z g = ZUtils.leastPrimitiveRoot(Z.valueOf(p2));
     final int q = g.modPow(Z.valueOf(mP), Z.valueOf(p2)).intValueExact();
     final CR[] etaStar = new CR[mP];
