+++ src/irvine/oeis/a018/A018937.java
@@ -32,7 +32,7 @@ public class A018937 extends Sequence1 {
   @Override
   public Z next() {
     ++mN;
-    mSumSquares += mN * mN;
+    mSumSquares += (long) mN * mN;
     long m = LongUtils.sqrt(mSumSquares);
     while (true) {
       final long soln = leastSolution(m * m, mN, 0);
